honeypot-links/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.31
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
qrcode==7.4.2
pillow==10.3.0
requests==2.32.3
